[apps/interactive_curve_view_range] Check preferences for autozoom memoization, to reset autozoom when angle unit or complex format are changed

+++ apps/shared/interactive_curve_view_range.cpp
@@ -347,7 +347,7 @@ void InteractiveCurveViewRange::privateComputeRanges(bool computeX,
       CircuitBreakerCheckpoint checkpoint(
           Ion::CircuitBreaker::CheckpointType::Back);
       if (CircuitBreakerRun(checkpoint)) {
-        uint32_t checksum;
+        uint64_t checksum;
         if (useMemoizedAutoRange &&
             (checksum = m_delegate->autoZoomChecksum()) ==
                 m_checksumOfMemoizedAutoRange) {
